Make test echo file.
