+++ tests/test_utils.py
@@ -1,16 +1,8 @@
 import os
 import pathlib
 import stat
-import subprocess
-import sys
-from collections import namedtuple
-from contextlib import nullcontext
 from decimal import Decimal
 from fractions import Fraction
-from functools import partial
-from io import BytesIO
-from io import StringIO
-from unittest.mock import patch
 
 import pytest
 
@@ -66,49 +58,6 @@ def test_unset_sentinel():
         assert value is not real_value
 
     assert value not in real_values
-
-
-def test_echo(runner):
-    with runner.isolation() as outstreams:
-        click.echo("\N{SNOWMAN}")
-        click.echo(b"\x44\x44")
-        click.echo(42, nl=False)
-        click.echo(b"a", nl=False)
-        click.echo("\x1b[31mx\x1b[39m", nl=False)
-        bytes = outstreams[0].getvalue().replace(b"\r\n", b"\n")
-        assert bytes == b"\xe2\x98\x83\nDD\n42ax"
-
-    # if wrapped, we expect bytes to survive.
-    @click.command()
-    def cli():
-        click.echo(b"\xf6")
-
-    result = runner.invoke(cli, [])
-    assert result.stdout_bytes == b"\xf6\n"
-
-    # Ensure we do not strip for bytes.
-    with runner.isolation() as outstreams:
-        click.echo(bytearray(b"\x1b[31mx\x1b[39m"), nl=False)
-        assert outstreams[0].getvalue() == b"\x1b[31mx\x1b[39m"
-
-
-def test_echo_custom_file():
-    f = StringIO()
-    click.echo("hello", file=f)
-    assert f.getvalue() == "hello\n"
-
-    b = BytesIO()
-    click.echo(b"", b)
-    assert b.getvalue() == b"\n"
-
-
-def test_echo_no_streams(monkeypatch, runner):
-    """echo should not fail when stdout and stderr are None with pythonw on Windows."""
-    with runner.isolation():
-        sys.stdout = None
-        sys.stderr = None
-        click.echo("test")
-        click.echo("test", err=True)
 
 
 @pytest.mark.parametrize(
@@ -275,385 +224,6 @@ def test_prompts_eof(runner):
     # only provide two lines of input for three prompts
     result = runner.invoke(echo, input="1\n2\n")
     assert result.exit_code == 1
-
-
-def _test_gen_func():
-    yield "a"
-    yield "b"
-    yield "c"
-    yield "abc"
-
-
-def _test_gen_func_fails():
-    raise RuntimeError("This is a test.")
-    yield  # unreachable, keeps this a generator function
-
-
-def _test_gen_func_yields_then_fails():
-    yield "test"
-    raise RuntimeError("This is a test.")
-
-
-def _test_gen_func_echo(file=None):
-    yield "test"
-    click.echo("hello", file=file)
-    yield "test"
-
-
-def _test_simulate_keyboard_interrupt(file=None):
-    yield "output_before_keyboard_interrupt"
-    raise KeyboardInterrupt()
-
-
-EchoViaPagerTest = namedtuple(
-    "EchoViaPagerTest",
-    (
-        "description",
-        "test_input",
-        "expected_pager",
-        "expected_stdout",
-        "expected_stderr",
-        "expected_error",
-    ),
-)
-
-
-@pytest.mark.skipif(WIN, reason="Different behavior on windows.")
-@pytest.mark.parametrize(
-    "pager_cmd", ["cat", "cat ", " cat ", "less", " less", " less "]
-)
-@pytest.mark.parametrize(
-    "test",
-    [
-        # We need to pass a parameter function instead of a plain param
-        # as pytest.mark.parametrize will reuse the parameters causing the
-        # generators to be used up so they will not yield anymore
-        EchoViaPagerTest(
-            description="Plain string argument",
-            test_input=lambda: "just text",
-            expected_pager="just text\n",
-            expected_stdout="",
-            expected_stderr="",
-            expected_error=None,
-        ),
-        EchoViaPagerTest(
-            description="Iterable argument",
-            test_input=lambda: ["itera", "ble"],
-            expected_pager="iterable\n",
-            expected_stdout="",
-            expected_stderr="",
-            expected_error=None,
-        ),
-        EchoViaPagerTest(
-            description="Generator function argument",
-            test_input=lambda: _test_gen_func,
-            expected_pager="abcabc\n",
-            expected_stdout="",
-            expected_stderr="",
-            expected_error=None,
-        ),
-        EchoViaPagerTest(
-            description="String generator argument",
-            test_input=lambda: _test_gen_func(),
-            expected_pager="abcabc\n",
-            expected_stdout="",
-            expected_stderr="",
-            expected_error=None,
-        ),
-        EchoViaPagerTest(
-            description="Number generator expression argument",
-            test_input=lambda: (c for c in range(6)),
-            expected_pager="012345\n",
-            expected_stdout="",
-            expected_stderr="",
-            expected_error=None,
-        ),
-        EchoViaPagerTest(
-            description="Exception in generator function argument",
-            test_input=lambda: _test_gen_func_fails,
-            # Because generator throws early on, the pager did not have
-            # a chance yet to write the file.
-            expected_pager="",
-            expected_stdout="",
-            expected_stderr="",
-            expected_error=RuntimeError,
-        ),
-        EchoViaPagerTest(
-            description="Exception in generator argument",
-            test_input=lambda: _test_gen_func_fails,
-            # Because generator throws early on, the pager did not have a
-            # chance yet to write the file.
-            expected_pager="",
-            expected_stdout="",
-            expected_stderr="",
-            expected_error=RuntimeError,
-        ),
-        EchoViaPagerTest(
-            description="Keyboard interrupt should not terminate the pager",
-            test_input=lambda: _test_simulate_keyboard_interrupt(),
-            # Due to the keyboard interrupt during pager execution, click program
-            # should abort, but the pager should stay open.
-            # This allows users to cancel the program and search in the pager
-            # output, before they decide to terminate the pager.
-            expected_pager="output_before_keyboard_interrupt",
-            expected_stdout="",
-            expected_stderr="",
-            expected_error=KeyboardInterrupt,
-        ),
-        EchoViaPagerTest(
-            description="Writing to stdout during generator execution",
-            test_input=lambda: _test_gen_func_echo(),
-            expected_pager="testtest\n",
-            expected_stdout="hello\n",
-            expected_stderr="",
-            expected_error=None,
-        ),
-        EchoViaPagerTest(
-            description="Writing to stderr during generator execution",
-            test_input=lambda: _test_gen_func_echo(file=sys.stderr),
-            expected_pager="testtest\n",
-            expected_stdout="",
-            expected_stderr="hello\n",
-            expected_error=None,
-        ),
-    ],
-)
-def test_echo_via_pager(monkeypatch, capfd, pager_cmd, test, tmp_path):
-    monkeypatch.setitem(os.environ, "PAGER", pager_cmd)
-    monkeypatch.setattr(click._termui_impl, "isatty", lambda x: True)
-
-    test_input = test.test_input()
-    expected_pager = test.expected_pager
-    expected_stdout = test.expected_stdout
-    expected_stderr = test.expected_stderr
-    expected_error = test.expected_error
-
-    check_raise = pytest.raises(expected_error) if expected_error else nullcontext()
-
-    pager_out_tmp = tmp_path / "pager_out.txt"
-    with pager_out_tmp.open("w") as f:
-        force_subprocess_stdout = patch.object(
-            subprocess,
-            "Popen",
-            partial(subprocess.Popen, stdout=f),
-        )
-        with force_subprocess_stdout:
-            with check_raise:
-                click.echo_via_pager(test_input)
-
-    out, err = capfd.readouterr()
-
-    pager = pager_out_tmp.read_text()
-
-    assert pager == expected_pager, (
-        f"Unexpected pager output in test case '{test.description}'"
-    )
-    assert out == expected_stdout, (
-        f"Unexpected stdout in test case '{test.description}'"
-    )
-    assert err == expected_stderr, (
-        f"Unexpected stderr in test case '{test.description}'"
-    )
-
-
-@pytest.mark.skipif(WIN, reason="Different behavior on windows.")
-def test_echo_via_pager_yields_before_exception(monkeypatch, tmp_path):
-    """A generator that yields then raises: click writes the partial output to
-    the pager stream before propagating the exception.
-
-    The pager file content is intentionally NOT asserted: pipe-drain timing
-    between click and the pager subprocess is outside click's control
-    (#2899, #3470). Spying on ``MaybeStripAnsi.write`` records what click sent
-    to the pager, which is deterministic regardless of scheduling.
-    """
-    monkeypatch.setitem(os.environ, "PAGER", "cat")
-    monkeypatch.setattr(click._termui_impl, "isatty", lambda x: True)
-
-    writes: list[str] = []
-    real_write = click._termui_impl.MaybeStripAnsi.write
-
-    def spy(self, text):
-        writes.append(text)
-        return real_write(self, text)
-
-    monkeypatch.setattr(click._termui_impl.MaybeStripAnsi, "write", spy)
-
-    pager_out_tmp = tmp_path / "pager_out.txt"
-    with (
-        pager_out_tmp.open("w") as f,
-        patch.object(subprocess, "Popen", partial(subprocess.Popen, stdout=f)),
-        pytest.raises(RuntimeError, match="This is a test."),
-    ):
-        click.echo_via_pager(_test_gen_func_yields_then_fails())
-
-    assert "".join(writes) == "test", (
-        f"click should have written the yielded chunk before exception, got {writes!r}"
-    )
-
-
-@pytest.mark.stress
-@pytest.mark.skipif(WIN, reason="Different behavior on windows.")
-@pytest.mark.parametrize("_", range(1000))
-def test_stress_echo_via_pager_exception_cleanup(_, monkeypatch, tmp_path):
-    """Repeated exceptions during ``echo_via_pager`` must not leak subprocesses.
-
-    Regression coverage for the cleanup path in ``_pipepager``'s exception
-    handler (issue #2899, PR #3470). Each iteration spawns a real pager
-    subprocess, raises before any data is written and check there is no leak.
-    """
-    monkeypatch.setitem(os.environ, "PAGER", "cat")
-    monkeypatch.setattr(click._termui_impl, "isatty", lambda x: True)
-
-    spawned: list[subprocess.Popen] = []
-    real_popen = subprocess.Popen
-
-    def tracking_popen(*args, **kwargs):
-        p = real_popen(*args, **kwargs)
-        spawned.append(p)
-        return p
-
-    pager_out_tmp = tmp_path / "pager_out.txt"
-    with (
-        pager_out_tmp.open("w") as f,
-        patch.object(subprocess, "Popen", partial(tracking_popen, stdout=f)),
-        pytest.raises(RuntimeError),
-    ):
-        click.echo_via_pager(_test_gen_func_fails())
-
-    assert spawned, "pager subprocess was never started"
-    for p in spawned:
-        assert p.returncode is not None, "pager subprocess not reaped"
-
-
-def test_echo_color_flag(monkeypatch, capfd):
-    isatty = True
-    monkeypatch.setattr(click._compat, "isatty", lambda x: isatty)
-
-    text = "foo"
-    styled_text = click.style(text, fg="red")
-    assert styled_text == "\x1b[31mfoo\x1b[0m"
-
-    click.echo(styled_text, color=False)
-    out, err = capfd.readouterr()
-    assert out == f"{text}\n"
-
-    click.echo(styled_text, color=True)
-    out, err = capfd.readouterr()
-    assert out == f"{styled_text}\n"
-
-    isatty = True
-    click.echo(styled_text)
-    out, err = capfd.readouterr()
-    assert out == f"{styled_text}\n"
-
-    isatty = False
-    # Faking isatty() is not enough on Windows;
-    # the implementation caches the colorama wrapped stream
-    # so we have to use a new stream for each test
-    stream = StringIO()
-    click.echo(styled_text, file=stream)
-    assert stream.getvalue() == f"{text}\n"
-
-    stream = StringIO()
-    click.echo(styled_text, file=stream, color=True)
-    assert stream.getvalue() == f"{styled_text}\n"
-
-
-def test_prompt_cast_default(capfd, monkeypatch):
-    monkeypatch.setattr(sys, "stdin", StringIO("\n"))
-    value = click.prompt("value", default="100", type=int)
-    capfd.readouterr()
-    assert isinstance(value, int)
-
-
-@pytest.mark.skipif(WIN, reason="Test too complex to make work windows.")
-def test_echo_writing_to_standard_error(capfd, monkeypatch):
-    def emulate_input(text):
-        """Emulate keyboard input."""
-        monkeypatch.setattr(sys, "stdin", StringIO(text))
-
-    click.echo("Echo to standard output")
-    out, err = capfd.readouterr()
-    assert out == "Echo to standard output\n"
-    assert err == ""
-
-    click.echo("Echo to standard error", err=True)
-    out, err = capfd.readouterr()
-    assert out == ""
-    assert err == "Echo to standard error\n"
-
-    emulate_input("asdlkj\n")
-    click.prompt("Prompt to stdin")
-    out, err = capfd.readouterr()
-    assert out == "Prompt to stdin: "
-    assert err == ""
-
-    emulate_input("asdlkj\n")
-    click.prompt("Prompt to stdin with no suffix", prompt_suffix="")
-    out, err = capfd.readouterr()
-    assert out == "Prompt to stdin with no suffix"
-    assert err == ""
-
-    # On non-Windows the full prompt goes through redirect_stdout so
-    # nothing leaks to stdout when err=True.
-    # https://github.com/pallets/click/issues/2968
-    emulate_input("asdlkj\n")
-    click.prompt("Prompt to stderr", err=True)
-    out, err = capfd.readouterr()
-    assert out == ""
-    assert err == "Prompt to stderr: "
-
-    # https://github.com/pallets/click/issues/3019
-    emulate_input("asdlkj\n")
-    click.prompt("Prompt to stderr with no suffix", prompt_suffix="", err=True)
-    out, err = capfd.readouterr()
-    assert out == ""
-    assert err == "Prompt to stderr with no suffix"
-
-    emulate_input("y\n")
-    click.confirm("Prompt to stdin")
-    out, err = capfd.readouterr()
-    assert out == "Prompt to stdin [y/N]: "
-    assert err == ""
-
-    emulate_input("y\n")
-    click.confirm("Prompt to stdin with no suffix", prompt_suffix="")
-    out, err = capfd.readouterr()
-    assert out == "Prompt to stdin with no suffix [y/N]"
-    assert err == ""
-
-    # https://github.com/pallets/click/issues/2968
-    emulate_input("y\n")
-    click.confirm("Prompt to stderr", err=True)
-    out, err = capfd.readouterr()
-    assert out == ""
-    assert err == "Prompt to stderr [y/N]: "
-
-    # https://github.com/pallets/click/issues/3019
-    emulate_input("y\n")
-    click.confirm("Prompt to stderr with no suffix", prompt_suffix="", err=True)
-    out, err = capfd.readouterr()
-    assert out == ""
-    assert err == "Prompt to stderr with no suffix [y/N]"
-
-    monkeypatch.setattr(click.termui, "isatty", lambda x: True)
-    monkeypatch.setattr(click.termui, "getchar", lambda: " ")
-
-    click.pause("Pause to stdout")
-    out, err = capfd.readouterr()
-    assert out == "Pause to stdout\n"
-    assert err == ""
-
-    click.pause("Pause to stderr", err=True)
-    out, err = capfd.readouterr()
-    assert out == ""
-    assert err == "Pause to stderr\n"
-
-
-def test_echo_with_capsys(capsys):
-    click.echo("Capture me.")
-    out, err = capsys.readouterr()
-    assert out == "Capture me.\n"
 
 
 def test_open_file(runner):
